+++ ops/framework.py
@@ -263,11 +263,11 @@ class EventSource:
 
     It is generally used as::
 
-        class SomethingHappened(EventBase):
+        class SomethingHappened(ops.EventBase):
             pass
 
         class SomeObject(Object):
-            something_happened = EventSource(SomethingHappened)
+            something_happened = ops.EventSource(SomethingHappened)
 
     With that, instances of that type will offer the ``someobj.something_happened``
     attribute which is a :class:`BoundEvent`, and may be used to emit and observe
@@ -331,7 +331,18 @@ class BoundEvent:
     def emit(self, *args: Any, **kwargs: Any):
         """Emit event to all registered observers.
 
-        The current storage state is committed before and after each observer is notified.
+        The current storage state is committed before and after each observer
+        is notified.
+
+        Note that the emission of custom events is handled immediately. In
+        other words, custom events are not queued, but rather nested. For
+        example::
+
+            1. Main hook handler (emits custom_event_1)
+            2.   Custom event 1 handler (emits custom_event_2)
+            3.     Custom event 2 handler
+            4.   Resume custom event 1 handler
+            5. Resume main hook handler
         """
         framework = self.emitter.framework
         key = framework._next_event_key()
@@ -445,6 +456,10 @@ class ObjectEvents(Object):
                         event. Must be a valid Python identifier, not be a keyword
                         or an existing attribute.
             event_type: A type of the event to define.
+
+        Raises:
+            RuntimeError: if the same event is defined twice, or if ``event_kind``
+                is an invalid name.
         """
         prefix = 'unable to define an event with event_kind that '
         if not event_kind.isidentifier():
@@ -710,7 +725,7 @@ class Framework(Object):
             marshal.dumps(data)
         except ValueError:
             msg = "unable to save the data for {}, it must contain only simple types: {!r}"
-            raise ValueError(msg.format(value.__class__.__name__, data))
+            raise ValueError(msg.format(value.__class__.__name__, data)) from None
 
         self._storage.save_snapshot(value.handle.path, data)
 
@@ -954,6 +969,9 @@ class Framework(Object):
         stop execution when a hook event is about to be handled.
 
         For those reasons, the "all" and "hook" breakpoint names are reserved.
+
+        Raises:
+            ValueError: if the breakpoint name is invalid.
         """
         # If given, validate the name comply with all the rules
         if name is not None:
@@ -1099,15 +1117,15 @@ class StoredState:
 
     Example::
 
-        class MyClass(Object):
-            _stored = StoredState()
+        class MyClass(ops.Object):
+            _stored = ops.StoredState()
 
     Instances of ``MyClass`` can transparently save state between invocations by
     setting attributes on ``_stored``. Initial state should be set with
     ``set_default`` on the bound object, that is::
 
-        class MyClass(Object):
-            _stored = StoredState()
+        class MyClass(ops.Object):
+            _stored = ops.StoredState()
 
             def __init__(self, parent, key):
                 super().__init__(parent, key)