+++ ops/pebble.py
@@ -125,16 +125,27 @@ CheckDict = typing.TypedDict('CheckDict',
                               'threshold': Optional[int]},
                              total=False)
 
+# In Python 3.11+ 'services' and 'labels' should be NotRequired, and total=True.
+LogTargetDict = typing.TypedDict('LogTargetDict',
+                                 {'override': Union[Literal['merge'], Literal['replace']],
+                                  'type': Literal['loki'],
+                                  'location': str,
+                                  'services': List[str],
+                                  'labels': Dict[str, str]},
+                                 total=False)
+
 LayerDict = typing.TypedDict('LayerDict',
                              {'summary': str,
                               'description': str,
                               'services': Dict[str, ServiceDict],
-                              'checks': Dict[str, CheckDict]},
+                              'checks': Dict[str, CheckDict],
+                              'log-targets': Dict[str, LogTargetDict]},
                              total=False)
 
 PlanDict = typing.TypedDict('PlanDict',
                             {'services': Dict[str, ServiceDict],
-                             'checks': Dict[str, CheckDict]},
+                             'checks': Dict[str, CheckDict],
+                             'log-targets': Dict[str, LogTargetDict]},
                             total=False)
 
 _AuthDict = TypedDict('_AuthDict',
@@ -734,6 +745,9 @@ class Plan:
                                               for name, service in d.get('services', {}).items()}
         self._checks: Dict[str, Check] = {name: Check(name, check)
                                           for name, check in d.get('checks', {}).items()}
+        self._log_targets: Dict[str, LogTarget] = {
+            name: LogTarget(name, target)
+            for name, target in d.get('log-targets', {}).items()}
 
     @property
     def services(self) -> Dict[str, 'Service']:
@@ -751,11 +765,20 @@ class Plan:
         """
         return self._checks
 
+    @property
+    def log_targets(self) -> Dict[str, 'LogTarget']:
+        """This plan's log targets mapping (maps log target name to :class:`LogTarget`).
+
+        This property is currently read-only.
+        """
+        return self._log_targets
+
     def to_dict(self) -> 'PlanDict':
         """Convert this plan to its dict representation."""
         fields = [
             ('services', {name: service.to_dict() for name, service in self._services.items()}),
             ('checks', {name: check.to_dict() for name, check in self._checks.items()}),
+            ('log-targets', {name: target.to_dict() for name, target in self._log_targets.items()})
         ]
         dct = {name: value for name, value in fields if value}
         return typing.cast('PlanDict', dct)
@@ -782,6 +805,8 @@ class Layer:
     services: Dict[str, 'Service']
     #: Mapping of check to :class:`Check` defined by this layer.
     checks: Dict[str, 'Check']
+    #: Mapping of target to :class:`LogTarget` defined by this layer.
+    log_targets: Dict[str, 'LogTarget']
 
     def __init__(self, raw: Optional[Union[str, 'LayerDict']] = None):
         if isinstance(raw, str):
@@ -796,6 +821,8 @@ class Layer:
                          for name, service in d.get('services', {}).items()}
         self.checks = {name: Check(name, check)
                        for name, check in d.get('checks', {}).items()}
+        self.log_targets = {name: LogTarget(name, target)
+                            for name, target in d.get('log-targets', {}).items()}
 
     def to_yaml(self) -> str:
         """Convert this layer to its YAML representation."""
@@ -808,6 +835,7 @@ class Layer:
             ('description', self.description),
             ('services', {name: service.to_dict() for name, service in self.services.items()}),
             ('checks', {name: check.to_dict() for name, check in self.checks.items()}),
+            ('log-targets', {name: target.to_dict() for name, target in self.log_targets.items()})
         ]
         dct = {name: value for name, value in fields if value}
         return typing.cast('LayerDict', dct)
@@ -1042,6 +1070,45 @@ class CheckStatus(enum.Enum):
 
     UP = 'up'
     DOWN = 'down'
+
+
+class LogTarget:
+    """Represents a log target in a Pebble configuration layer."""
+
+    def __init__(self, name: str, raw: Optional['LogTargetDict'] = None):
+        self.name = name
+        dct: LogTargetDict = raw or {}
+        self.override: str = dct.get('override', '')
+        self.type = dct.get('type', '')
+        self.location = dct.get('location', '')
+        self.services: List[str] = list(dct.get('services', []))
+        labels = dct.get('labels')
+        if labels is not None:
+            labels = copy.deepcopy(labels)
+        self.labels: Optional[Dict[str, str]] = labels
+
+    def to_dict(self) -> 'LogTargetDict':
+        """Convert this log target object to its dict representation."""
+        fields = [
+            ('override', self.override),
+            ('type', self.type),
+            ('location', self.location),
+            ('services', self.services),
+            ('labels', self.labels),
+        ]
+        dct = {name: value for name, value in fields if value}
+        return typing.cast('LogTargetDict', dct)
+
+    def __repr__(self):
+        return f'LogTarget({self.to_dict()!r})'
+
+    def __eq__(self, other: Union['LogTargetDict', 'LogTarget']):
+        if isinstance(other, dict):
+            return self.to_dict() == other
+        elif isinstance(other, LogTarget):
+            return self.to_dict() == other.to_dict()
+        else:
+            return NotImplemented
 
 
 class FileType(enum.Enum):
@@ -1452,6 +1519,7 @@ class ExecProcess(Generic[AnyStr]):
         Raises:
             ChangeError: if there was an error starting or running the process.
             ExecError: if the process exits with a non-zero exit code.
+            TypeError: if :meth:`Client.exec` was called with the ``stdout`` argument.
         """
         if self.stdout is None:
             raise TypeError(
@@ -1636,6 +1704,15 @@ class Client:
 
     Defaults to using a Unix socket at socket_path (which must be specified
     unless a custom opener is provided).
+
+    For methods that wait for changes, such as :meth:`start_services` and :meth:`replan_services`,
+    if the change fails or times out, then a :class:`ChangeError` or :class:`TimeoutError` will be
+    raised.
+
+    All methods may raise exceptions when there are problems communicating with Pebble. Problems
+    connecting to or transferring data with Pebble will raise a :class:`ConnectionError`. When an
+    error occurs executing the request, such as trying to add an invalid layer or execute a command
+    that does not exist, an :class:`APIError` is raised.
     """
 
     _chunk_size = 8192
@@ -1729,9 +1806,13 @@ class Client:
                 # Will only happen on read error or if Pebble sends invalid JSON.
                 body: Dict[str, Any] = {}
                 message = f'{type(e2).__name__} - {e2}'
-            raise APIError(body, code, status, message)
+            raise APIError(body, code, status, message) from None
         except urllib.error.URLError as e:
-            raise ConnectionError(e.reason)
+            if e.args and isinstance(e.args[0], FileNotFoundError):
+                raise ConnectionError(
+                    f"Could not connect to Pebble: socket not found at {self.socket_path!r} "
+                    "(container restarted?)") from None
+            raise ConnectionError(e.reason) from e
 
         return response
 
@@ -1777,16 +1858,16 @@ class Client:
         """Start the startup-enabled services and wait (poll) for them to be started.
 
         Args:
-            timeout: Seconds before autostart change is considered timed out (float).
+            timeout: Seconds before autostart change is considered timed out (float). If
+                timeout is 0, submit the action but don't wait; just return the change ID
+                immediately.
             delay: Seconds before executing the autostart change (float).
 
         Returns:
             ChangeID of the autostart change.
 
         Raises:
-            ChangeError: if one or more of the services didn't start. If
-                timeout is 0, submit the action but don't wait; just return the change
-                ID immediately.
+            ChangeError: if one or more of the services didn't start, and ``timeout`` is non-zero.
         """
         return self._services_action('autostart', [], timeout, delay)
 
@@ -1794,16 +1875,17 @@ class Client:
         """Replan by (re)starting changed and startup-enabled services and wait for them to start.
 
         Args:
-            timeout: Seconds before replan change is considered timed out (float).
+            timeout: Seconds before replan change is considered timed out (float). If
+                timeout is 0, submit the action but don't wait; just return the change
+                ID immediately.
             delay: Seconds before executing the replan change (float).
 
         Returns:
             ChangeID of the replan change.
 
         Raises:
-            ChangeError: if one or more of the services didn't stop/start. If
-                timeout is 0, submit the action but don't wait; just return the change
-                ID immediately.
+            ChangeError: if one or more of the services didn't stop/start, and ``timeout`` is
+                non-zero.
         """
         return self._services_action('replan', [], timeout, delay)
 
@@ -1814,16 +1896,17 @@ class Client:
 
         Args:
             services: Non-empty list of services to start.
-            timeout: Seconds before start change is considered timed out (float).
+            timeout: Seconds before start change is considered timed out (float). If
+                timeout is 0, submit the action but don't wait; just return the change
+                ID immediately.
             delay: Seconds before executing the start change (float).
 
         Returns:
             ChangeID of the start change.
 
         Raises:
-            ChangeError: if one or more of the services didn't stop/start. If
-                timeout is 0, submit the action but don't wait; just return the change
-                ID immediately.
+            ChangeError: if one or more of the services didn't stop/start, and ``timeout`` is
+                non-zero.
         """
         return self._services_action('start', services, timeout, delay)
 
@@ -1834,16 +1917,17 @@ class Client:
 
         Args:
             services: Non-empty list of services to stop.
-            timeout: Seconds before stop change is considered timed out (float).
+            timeout: Seconds before stop change is considered timed out (float). If
+                timeout is 0, submit the action but don't wait; just return the change
+                ID immediately.
             delay: Seconds before executing the stop change (float).
 
         Returns:
             ChangeID of the stop change.
 
         Raises:
-            ChangeError: if one or more of the services didn't stop/start. If
-                timeout is 0, submit the action but don't wait; just return the change
-                ID immediately.
+            ChangeError: if one or more of the services didn't stop/start and ``timeout`` is
+                non-zero.
         """
         return self._services_action('stop', services, timeout, delay)
 
@@ -1854,16 +1938,17 @@ class Client:
 
         Args:
             services: Non-empty list of services to restart.
-            timeout: Seconds before restart change is considered timed out (float).
+            timeout: Seconds before restart change is considered timed out (float). If
+                timeout is 0, submit the action but don't wait; just return the change
+                ID immediately.
             delay: Seconds before executing the restart change (float).
 
         Returns:
             ChangeID of the restart change.
 
         Raises:
-            ChangeError: if one or more of the services didn't stop/start. If
-                timeout is 0, submit the action but don't wait; just return the change
-                ID immediately.
+            ChangeError: if one or more of the services didn't stop/start and ``timeout`` is
+                non-zero.
         """
         return self._services_action('restart', services, timeout, delay)
 
@@ -2125,6 +2210,10 @@ class Client:
             group_id: Group ID (GID) for file.
             group: Group name for file. Group's GID must match group_id if
                 both are specified.
+
+        Raises:
+            PathError: If there was an error writing the file to the path; for example, if the
+                destination path doesn't exist and ``make_dirs`` is not used.
         """
         info = self._make_auth_dict(permissions, user_id, user, group_id, group)
         info['path'] = path
@@ -2222,6 +2311,10 @@ class Client:
                 for example ``*.txt``.
             itself: If path refers to a directory, return information about the
                 directory itself, rather than its contents.
+
+        Raises:
+            PathError: if there was an error listing the directory; for example, if the directory
+                does not exist.
         """
         query = {
             'action': 'list',
@@ -2255,6 +2348,10 @@ class Client:
             group_id: Group ID (GID) for directory.
             group: Group name for directory. Group's GID must match group_id
                 if both are specified.
+
+        Raises:
+            PathError: if there was an error making the directory; for example, if the parent path
+                does not exist, and ``make_parents`` is not used.
         """
         info = self._make_auth_dict(permissions, user_id, user, group_id, group)
         info['path'] = path
@@ -2280,7 +2377,6 @@ class Client:
         Raises:
             pebble.PathError: If a relative path is provided, or if `recursive` is False
                 and the file or directory cannot be removed (it does not exist or is not empty).
-
         """
         info: Dict[str, Any] = {'path': path}
         if recursive:
@@ -2472,6 +2568,11 @@ class Client:
             :meth:`ExecProcess.wait` if stdout/stderr were provided as
             arguments to :meth:`exec`, or :meth:`ExecProcess.wait_output` if
             not.
+
+        Raises:
+            APIError: if an error occurred communicating with pebble, or if the command is not
+                found.
+            ExecError: if the command exits with a non-zero exit code.
         """
         if not isinstance(command, list) or not all(isinstance(s, str) for s in command):
             raise TypeError(f'command must be a list of str, not {type(command).__name__}')