+++ test/test_log.py
@@ -1,5 +1,3 @@
-#!/usr/bin/python3
-
 # Copyright 2020 Canonical Ltd.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");