+++ CHANGES.md
@@ -1,6 +1,14 @@
+# 2.9.0
+
+* Added log target support to `ops.pebble` layers and plans.
+* Added `Harness.run_action()`, `testing.ActionOutput`, and `testing.ActionFailed`
+
 # 2.8.0
 
+* Added `Unit.reboot()` and `Harness.reboot_count``
+* Added `RelationMeta.optional`
 * The type of a `Handle`'s `key` was expanded from `str` to `str|None`
+* Narrowed types of `app` and `unit` in relation events to exclude `None` where applicable
 
 # 2.7.0